+++ evm-template/runtime/src/governance/mod.rs
@@ -12,7 +12,10 @@ use frame_system::EnsureRootWithSuccess;
 
 use super::*;
 use crate::{
-    constants::currency::{CENTS, GRAND},
+    constants::{
+        currency::{CENTS, GRAND},
+        DAYS,
+    },
     Balance,
 };
