+++ evm-template/runtime/src/governance/tracks.rs
@@ -1,6 +1,7 @@
 //! Track configurations for governance.
 
 use super::*;
+use crate::constants::MINUTES;
 
 const fn percent(x: i32) -> sp_arithmetic::FixedI64 {
     sp_arithmetic::FixedI64::from_rational(x as u128, 100)
